+++ fragments/step.glsl
@@ -4,7 +4,7 @@
 #define HALF_PI 1.57079632679
 
 
-precision highp float; // Good practice for precision
+precision mediump float; // Good practice for precision
 
 in vec2 texCoord; // Interpolated texture coordinate from the vertex shader
